src/requirements.txt
altgraph==0.17.3
autopep8==2.0.1
future==0.18.3
importlib-metadata==6.0.0
pefile==2022.5.30
pycodestyle==2.10.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
PyQt6==6.4.0
PyQt6-Qt6==6.4.2
PyQt6-sip==13.4.0
pywin32-ctypes==0.2.0
PyYAML==6.0
tomli==2.0.1
typing_extensions==4.4.0
zipp==3.11.0
